fixing routing issues
